// src/domains/dashboard/Payroll/types/salaryProfileTypes/employeeSalaryTable.ts
export type SalaryDetailsRow = {
    key: string;
    componentName: string;
    category: string;
    amount: number;
};
// ...
export type SalaryItem = SalaryDetailsRow & {
    amount: string;
};
// ...
export type salarytableType = {
    id: string;
    name: string;
    employeeId: string;
    role: string;
    basicSalary: string | number;
    others: string | number;
    totalSalary?: string;
    totalDeduction?: string | number;
    status: string;
    action: string;
    email: string;
    image: string;
    department: string;
    salaryId: string;
    eId: string;
    employeeStatus?: string;
    lastWorkingDay?: string;
    monthlySalary?: string | number;
    totalPayable?: string | number;
    tds?: string | number;
    netSalaryBeforeTax?: string | number;
    oneTimeAmount?: number;
};

export interface SalaryCycleType {
    salaryCycleStart: string;
    salaryCycleEnd: string;
    salaryCycleDays: number;
}
export type filterState = {
    searchText: string;
    sort: string;
    page: number;
    filter: string;
    year: number;
    month: string | number;
    limit: number;
};
// ...
export type SalaryInfo = {
    basicPay?: number;
    hraAmount?: number;
    daAmount?: number;
    bonus?: number;
    incentiveAmount?: number;
    increamentAmount?: number;
    overtimeAmount?: number;
    deductionAmount?: number;
    leavesAmount?: number;
    tdsAmount?: number;
    other?: number;
};